pyproject.toml
[project]
name = "nextcloud-notes-mcp"
version = "0.1.1"
description = "Nextcloud Notes MCP Server"
readme = "README.md"
authors = [{ name = "ai_rncz" }]
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.3.0",
  "webdavclient3",
  "python-dotenv"
  ]

[project.scripts]
nextcloud-notes-mcp = "nextcloud_notes_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/rncz/nextcloud-notes-mcp-server"


